Fix bugs in rgl and rglh aliases
Before, the "sort" option was passed after the user's options. This led to
issues if you tried to pass "--" as an argument (e.g in order to search for a
string that started with a dash). Now, all of the alias' arguments (including
sort) come before all of the user's arguments.