Extract .ssh/config into separate files
Instead of one large .ssh/config, the file is now extracted into multiple files
in the .ssh/configs/ folder. There is a function, `update_ssh`, that combines
the files into the traditional .ssh/config file tools expect.

The reason behind this is to allow there to be a local, untracked SSH config
file (for work, etc.) that doesn't come along with the whole repository.