Improve yesno function and add noyes
A common functionality provided by command line scripts is to be able
to choose yes or no with one of the two being the default if you press
enter without entering anything.

Now, the `yesno` function will choose "yes" if the user presses enter
without choosing anything (i.e. enters an empty string). In addition,
a `noyes` function is provided. It performs the same function but
opposite: if the user does _not_ press anything, then "no" is
selected.