Use git push -u by default
The `-u` flag to `git push` means that Git will automatically tracks
the remote branch when you push a branch. This is what I want when I
create branches from the command line, push them, and later want to
pull them. Without it, I would manually have to specify which remote
branch it should track.

For branches that I don't create, this is irrelevant. Switching to
(checking out) a branch already automatically tracks the upstream
branch.