Improve jirafyfile scripts
I'm not sure why I was replacing Java code blocks with `noformat`
blocks. I think it was because even if I didn't specify a language it
would default to Java?

Remove the improper backslash escape `&bsol;` with a regular
backslash.

Remove the trailing blank line that appears at the end of code
blocks.