Improve gf Bash Git fzf function
The `gf` function did not properly escape all special characters. This
was mostly due to the inconsistent nature of `git status --short`.

The `gf` function now properly escapes all special characters by
wrapping the file name in single quotes only if it needs to. It also
properly handles renamed files.

This made it much more complicated, but I have tested it against all
kinds of weird scenarios, like renaming files that have a `->` in
their names. It appears to work in every scenario.