git-switch-autostash-skipped
This script was for switching branches when I told Git to ignore my
changes using `git update-index --skip-worktree`. For the most part
this worked, but it posed problems when various branches had different
versions of the file. This script was the beginning of an effort to
make it easier to switch branches with modified "skipped" files. It
DOES NOT WORK but I wanted to save the effort I had put in. I no
longer need this tool.