Fix Git Bash completion on Ubuntu
On Ubuntu, completions are lazy-loaded. This is normally good, but the
`__git_complete` function is _also_ lazy loaded, which we would like
to use to add completions for various functions and aliases. We use
the `_completion_loader` function to force loading of the Git
completion script and provide the `__git_complete` function.