Add completion for functions that take commands as arguments
The `eachdir` and `eachfile` functions should be completed as if you
were beginning a new command. This is exactly the purpose of the
`_command` completion generator.

The `editbin`, `openbin`, and `cdbin` functions all take a single
executable as an argument. This is similar to generating completion
for a new command. This is technically suboptimal since it will
continue trying to complete for subsequent arguments if you press tab
again. However, the user normally won't press tab again, since the
commands only take a single argument.