Improve PATH definitions
The list of elements to add to the `$PATH` variable were in an array
called `paths_rev`. This implied that they needed to be
reversed. However, they actually needed to be listed in order--they
just needed to be inserted into the `$PATH` variable in reverse order
since they were always added to the beginning.

This swaps the variable names to make that more clear.