Improve f and fh aliases/functions
The `f` and `fh` aliases are used to open files and directories in the
native macOS Finder application.

However, if you used `f` on a file, it would simply open the file with
the default application, much like the `open` command (as if you
double-clicked the file in the Finder application itself). If I want
to do that, I will use the `open` command. When I use the `f` command,
I want it to open in Finder. To support that, you can use `open -R`,
which "reveals" the file in Finder.

The new `f` function will use `open -R` for files and `open -a Finder
DIR` for directories. The `fh` function expects no arguments and opens
the current directory in Finder.