Improve gc command
The `gc` command uses Git and fzf to allow the user to select one or
more branches. However, in the past, even if the user selected
multiple branches, only one would be displayed.

Now, allow the user to select multiple branches and actually display
all the branches the user selected.