Create zipdir
Zips the entire contents of a directory into a ZIP archive. Each file
in the directory is stored at the root level of the archive. The name
of the ZIP file is the name of the directory plus the ".zip" suffix.

This is useful when you want to create a ZIP file where every file is
at the root level but you don't want to `cd` into the directory every
time.