Support Bash 3 in dotfiles
The dotfiles script used to run fine with Bash 3 (the default Bash on
OSX). However, when I added support for custom file mappings, I used
associative arrays, which were only added in Bash 4.

Convert this code to use regular arrays and nested for loops to find
matches. This can now run on Bash 3.