Add is_wsl and is_windows methods
In addition to using Bash in Linux and MacOS, I now use Bash in
Windows in two different ways: "[Windows Subsystem for
Linux][wsl]" (WSL) and "Git BASH" provided by [Git for
Windows][git-for-windows] via [Mingw-w64][mingw-w64].

These allow me to provide custom implementations for certain
operations that are not provided by or need to be different in WSL or
Git Bash (e.g. sending text to the clipboard).

[wsl]: https://learn.microsoft.com/en-us/windows/wsl/about
[git-for-windows]: https://gitforwindows.org/
[mingw-w64]: https://www.mingw-w64.org/