Increase maximum file limit on MacOS
In the past, I have run the `ulimit -n` command in an attempt to
increase the maximum number of open files. However, what I was doing
could not ratchet the maximum number higher than the level configured
by `launchctl`.

Avoid this by using `launchctl` to drastically increase the maximum
number of open files.